fix: repeating transactions in TransferPassword
TransferPassword would send the same transaction as many times as enter
would be pressed. The keyboard was supposed to be released when the
component stops being rendered, but it was rendered more
than expected due to being inside a Transition.

The fix is to keep track if a transaction was sent already,
and reset the variable next time the component is used.